// models/CustomAttributeTypedArgument.ts
/**
 * GameTrip.API
 * GameTrip API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Type } from '../models/Type';
import { HttpFile } from '../http/http';

export class CustomAttributeTypedArgument {
    'argumentType'?: Type;
    'value'?: any | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "argumentType",
            "baseName": "argumentType",
            "type": "Type",
            "format": ""
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "any",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CustomAttributeTypedArgument.attributeTypeMap;
    }

    public constructor() {
    }
}