// models/Exception.ts
/**
 * GameTrip.API
 * GameTrip API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { MethodBase } from '../models/MethodBase';
import { HttpFile } from '../http/http';

export class Exception {
    'targetSite'?: MethodBase;
    'message'?: string | null;
    'data'?: { [key: string]: any; } | null;
    'innerException'?: Exception;
    'helpLink'?: string | null;
    'source'?: string | null;
    'hResult'?: number;
    'stackTrace'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "targetSite",
            "baseName": "targetSite",
            "type": "MethodBase",
            "format": ""
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "{ [key: string]: any; }",
            "format": ""
        },
        {
            "name": "innerException",
            "baseName": "innerException",
            "type": "Exception",
            "format": ""
        },
        {
            "name": "helpLink",
            "baseName": "helpLink",
            "type": "string",
            "format": ""
        },
        {
            "name": "source",
            "baseName": "source",
            "type": "string",
            "format": ""
        },
        {
            "name": "hResult",
            "baseName": "hResult",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "stackTrace",
            "baseName": "stackTrace",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Exception.attributeTypeMap;
    }

    public constructor() {
    }
}