// models/FieldAttributes.ts
/**
 * GameTrip.API
 * GameTrip API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export type FieldAttributes = "0" | "1" | "2" | "3" | "4" | "5" | "6" | "7" | "16" | "32" | "64" | "128" | "256" | "512" | "1024" | "4096" | "8192" | "32768" | "38144" ;